Began refactoring with some docstring help The goal is to put general ft260 lib wrappers in ft_function. Put high level wrappers in ft. Put logger in ft only. Put GUI in two different files (UART I2C) atop ft.
